requirements.txt
aiogram==2.20
pyDes==2.0.1
python-dotenv==0.20.0
